mnist-cnn/requirements.txt
torch==1.5.0+cpu
torchvision==0.6.0+cpu
neuralpy-torch==0.1.0